refactor(ac-daemon): split handlers.rs into per-concern modules
Closes #29.

The 1964-line handlers.rs grew into a junk drawer: dispatch logic,
audio commands, calibrate state machine, two test suites, math
helpers, and DMM I/O all in one file. Concerns are now in
handlers/mod.rs (shared helpers + macro + re-exports) plus six
submodules:

  admin.rs     — status / quit / stop / devices / setup / server_*
  audio.rs     — generate / sweep_* / plot* / monitor_spectrum
  calibrate.rs — calibrate state machine + cal_reply
  transfer.rs  — transfer + probe (port-routing aware)
  test_hw.rs   — test_hardware + hw_* helpers
  test_dut.rs  — test_dut + dut_* helpers + dut_reply

server.rs dispatch is unchanged — public functions are re-exported
from handlers/mod.rs so the call sites keep working as
handlers::status, handlers::generate, etc.

No behavior change. cargo test -p ac-daemon: 10 passing (3 unit + 7
integration). Zero new warnings.
